Fix bug where adult mobs on a breeding cooldown can be turned into babies.
When a mob is on a breeding cooldown, their age is set to a positive number which decrements until it reaches 0. Using NBT tags, we can only test for the existence (or nonexistence) of an NBT tag, but we need to check that the age is less than 0. To sidestep this, I added a scoreboard age which temporarily stores the age of nearby mobs. After comparing the age of mobs, the scoreboard is cleared. I also had to add setup and uninstall scripts to create and delete the scoreboard. The setup script is automatically added by the load tag.